добавляет модуляцию
